// src/app/interfaces/measurement-data-response.ts
export interface MeasurementData {
    id:         string;
    type:       string;
    properties: Properties;
    geometry:   Geometry;
}

export interface Geometry {
    type:        string;
    coordinates: number[];
}

export interface Properties {
    OBJECTID_left:                                          number;
    probability:                                            number;
// ...
}